data types, operations, round methons and string formating

+++ percentage.py
@@ -1,4 +1,4 @@
 name = input("Enter Your name: ")
-sales = float(input("enter your sales for the month: "))
+sales = int(input("enter your sales for the month: "))
 
 print(f"hey {name} you this months sales is ${sales} and your commision for the month is ${round(sales*13/100,2)}")
